Move SearchBar from Components to Pages
